Refactor to extract `mapProposals` helper
Both the current proposal list table and the new proposal list sidebar
want the same proposals formatted in the same way. Move the helper
function `mapProposals` out to its own file, along with the helper
function it relies on.

Issue #62 Add sidebar to Proposal Detail page
